regmap: Support IO protocols for accessing indirect memory views

Closes HTD-171

See merge request ht/regio!29

+++ src/regio/regmap/proxy/variable.py
@@ -1,6 +1,8 @@
 #---------------------------------------------------------------------------------------------------
 __all__ = ()
 
+import math
+
 from ..io import io
 from ..spec import address, array, register, structure, union
 
@@ -11,6 +13,7 @@ class Variable:
 
         self._node = node
         self._chain = chain
+        self.is_group = chain.is_group if chain is not None else False
         self._pargs = tuple(pargs)
         self._kargs = dict(kargs)
 
@@ -18,21 +21,57 @@ class Variable:
         # buffered copy of it (by value).
         if initializer is Ellipsis:
             # By reference.
+            self.is_buffered = False
             self._context = ctx
         else:
             # By value.
-            # If the context is already buffered, pull-out it's low-level IO for the buffering.
-            # TODO: Should be able to use a custom buffer based on array.array(), where the index
-            #       is the register ordinal, adjusted for the first register in the sub-tree (all
-            #       registers in the sub-tree will be in a contiguous ordinal range).
-            llio = ctx.io.llio if isinstance(ctx.io, io.BufferedIO) else ctx.io
-            self._context = ctx.copy(io.BufferedIO(llio))
+            self.is_buffered = True
+            self._context = self._new_buffered_context(ctx)
             self.load(initializer)
 
         # Setup a proxy for the variable on the initialized context.
         self.proxy = self._context.new_proxy(node, chain)
 
+    def _new_buffered_context(self, ctx):
+        # Determine the low-level IO accessor to be buffered.
+        # - If the context is unbuffered, then use it's IO directly.
+        # - If the context is already buffered, then pull-out it's low-level IO for the buffering
+        #   (there's no current use case for layering of buffers, since by design, a buffer needs
+        #   some way to perform actual IO for load/store operations).
+        llio = ctx.io.llio if isinstance(ctx.io, io.BufferedIO) else ctx.io
+
+        # Wrap the low-level IO to buffer it's accesses.
+        buff_io = io.BufferedIO(llio)
+        if llio.started:
+            buff_io.start()
+
+        # Create the new context for the buffered IO.
+        return ctx.copy(buff_io)
+
+    def _iter_nodes(self, node, reg_only=True, iter_fn=lambda _: iter):
+        # Iterator for recursing depth-first through a node's hierarchy.
+        for child in iter_fn(node)(node.children):
+            # Only produce non-register nodes when requested.
+            is_reg = child.region.register is not None
+            if not reg_only or is_reg:
+                yield child
+
+            # Don't recurse through a register's fields unless requested.
+            if reg_only and is_reg:
+                continue
+
+            # Don't recurse through nodes requiring indirect access via protocol.
+            if child.protocol is not None:
+                continue
+
+            # Continue the depth-first iteration on the child node.
+            for gchild in self._iter_nodes(child, reg_only, iter_fn):
+                yield gchild
+
     def load(self, initializer=None):
+        if not self.is_buffered:
+            return
+
         # Pass the default value down to the IO buffer for use during buffered reads.
         if isinstance(initializer, int):
             self._context.io.default = initializer
@@ -43,14 +82,10 @@ class Variable:
         if initializer is not None:
             raise ValueError(f'Unknown initializer {initializer!r}. Must be None or an int.')
 
-        # The variable was created on a singular node, load all registers in it's hierarchy.
-        if not self._chain.is_group:
-            self._load_node(self._node)
-            return
-
-        # The variable was created on a node group, load all registers in the hierarchy of every
-        # node in the group.
-        for node in self._chain:
+        # Load all directly accessible registers in the hierarchy of the singular node or all nodes
+        # in the chain.
+        nodes = self._chain if self.is_group else (self._node,)
+        for node in nodes:
             self._load_node(node)
 
     def _load_node(self, node):
@@ -59,16 +94,20 @@ class Variable:
 
         # The node is itself a register.
         if node.region.register is not None:
-            load_region(node.region)
+            if node.config.access.is_readable:
+                load_region(node.region)
             return
 
-        # The node is a container, so load all registers in it's hierarchy.
+        # The node is a container, so load all directly accessible registers in it's hierarchy.
         # TODO: Don't walk the whole sub-tree. No need to walk past a register node.
-        for child in node.descendants:
-            if child.region.register is not None:
+        for child in self._iter_nodes(node):
+            if child.config.access.is_readable:
                 load_region(child.region)
 
     def store(self, initializer=None):
+        if not self.is_buffered:
+            return
+
         # Write all buffered data.
         if initializer is None:
             self.sync()
@@ -79,15 +118,11 @@ class Variable:
         if not isinstance(initializer, int):
             raise ValueError(f'Unknown initializer {initializer!r}. Must be None or an int.')
 
-        # The variable was created on a singular node, store all registers in it's hierarchy.
-        if not self._chain.is_group:
-            self._store_node(self._node, initializer)
-            return
-
-        # The variable was created on a node group, store all registers in the hierarchy of every
-        # node in the group.
-        for node in self._chain:
-            self._store_node(node, initializer)
+        # Store all directly accessible registers in the hierarchy of the singular node or all nodes
+        # in the chain.
+        nodes = self._chain if self.is_group else (self._node,)
+        for node in nodes:
+            self._store_node(node)
 
     def _store_node(self, node, initializer):
         # Perform low-level IO to write all registers.
@@ -95,23 +130,27 @@ class Variable:
 
         # The node is itself a register.
         if node.region.register is not None:
-            store_region(node.region, initializer)
+            if node.config.access.is_writeable:
+                store_region(node.region, initializer)
             return
 
-        # The node is a container, so store all registers in it's hierarchy.
+        # The node is a container, so store all directly accessible registers in it's hierarchy.
         # TODO: Don't walk the whole sub-tree. No need to walk past a register node.
-        for child in node.descendants:
-            if child.region.register is not None:
+        for child in self._iter_nodes(node):
+            if child.config.access.is_writeable:
                 store_region(child.region, initializer)
 
     def sync(self):
-        self._context.io.sync()
+        if self.is_buffered:
+            self._context.io.sync()
 
     def drop(self):
-        self._context.io.drop()
+        if self.is_buffered:
+            self._context.io.drop()
 
     def flush(self):
-        self._context.io.flush()
+        if self.is_buffered:
+            self._context.io.flush()
 
     def config_get(self, key, default=None):
         for kargs in (self._kargs, self._context.kargs):
@@ -135,10 +174,10 @@ class Formatter:
     def __init__(self, var):
         self.var = var
         self.root = var._node
-        self.is_group = var._chain.is_group
 
         # Capture general display configuration.
         self.verbose = var.config_get('verbose', False)
+        self.indirect = self.verbose or var.config_get('indirect', False)
 
         # Determine the display format for the qualified name of all the nodes.
         self.abspath = var.config_get('abspath', False)
@@ -147,11 +186,9 @@ class Formatter:
         # Sort lexicographically or leave in the order defined in the regmap specification.
         self.path_sort = self.var.config_get('path_sort', False)
 
-        # Determine the maximum number of nibbles for consistent offset display.
-        region = var._node.region
-        self.offset_units = 'Bytes' # TODO: Get from low-level IO.
-        self.offset_scale = region.data_width // 8 # TODO: Get from low-level IO.
-        self.offset_nibbles = ((region.size * self.offset_scale).bit_length() + 4 - 1) // 4
+        # Setup the display units for offsets.
+        self.offset_units = 'Bytes' # TODO: Provide option for alternatives (i.e. in words).
+        self.units_width = 8 # TODO: Get from low-level IO.
 
         # Determine the formatting for values.
         self.ignore_access = var.config_get('ignore_access', False)
@@ -176,16 +213,27 @@ class Formatter:
     def qualname(self, node, is_root=True):
         return self.abs_qualname(node) if self.abspath else self.rel_qualname(node, is_root)
 
-    def size(self, value):
-        value *= self.offset_scale
+    def offset_scale(self, region):
+        return (region.data_width + self.units_width - 1) // self.units_width
+
+    def size(self, region):
+        value = region.size * self.offset_scale(region)
         return f'{value:,}'
 
-    def offset(self, value):
-        value *= self.offset_scale
-        return f'0x{value:0{self.offset_nibbles}x}'
+    def offset(self, region, value=None, addend=0):
+        scale = self.offset_scale(region)
+        nibbles = (math.ceil(math.log2(region.root.size * scale)) + 4 - 1) // 4
 
-    def offset_range(self, start, end):
-        return self.offset(start) + ' - ' + self.offset(end)
+        if value is None:
+            value = region.offset.absolute
+        value = value * scale + addend
+
+        return f'0x{value:0{nibbles}x}'
+
+    def offset_range(self, region):
+        start = self.offset(region)
+        end = self.offset(region, region.offset.absolute + region.size, -1)
+        return start + ' - ' + end
 
     def value_hex(self, value, region):
         if self.with_hex_grouping:
@@ -217,36 +265,42 @@ class Formatter:
         return data
 
     def __str__(self):
-        root = self.root
-        ctx = self.var._context
-        chain = self.var._chain
+        # Define helper function to switch iteration order for register fields.
+        def iter_fn(node):
+            if isinstance(node, register.Node) and not self.lsb_first:
+                return reversed
+            return iter
+
+        col_widths = {}
+        row_data = []
+        def format_node(node, var, is_root):
+            ctx = var._context
+
+            # Format the node.
+            if is_root:
+                col_data = ctx.new_variable(node, None, ...)._format_node(self, True)
+                row_data.append(self.update_widths(col_widths, col_data))
+
+            # Gather formatting data for each child node in the hierarchy.
+            for child in var._iter_nodes(node, False, iter_fn):
+                child_ctx = ctx.select(child, None) if self.indirect else ctx
+                is_indirect = child_ctx is not ctx
+
+                child_var = child_ctx.new_variable(child, None, None if is_indirect else ...)
+                col_data = child_var._format_node(self, False)
+                row_data.append(self.update_widths(col_widths, col_data))
+
+                if is_indirect:
+                    format_node(child, child_var, False)
 
         # Determine the set of nodes to be formatted.
         # - If the variable was created on a node group, iterate over all nodes in the router chain.
         # - If the variable was created on a singular node, use it.
-        nodes = chain if chain.is_group else (root,)
-
-        # Setup an iterator for recursing depth-first through the regmap hierarchy.
-        def iter_nodes(node):
-            iter_fn = reversed if isinstance(node, register.Node) and not self.lsb_first else iter
-            for child in iter_fn(node.children):
-                yield child
-
-                for gchild in iter_nodes(child):
-                    yield gchild
+        nodes = self.var._chain if self.var.is_group else (self.root,)
 
         # Gather formatting data for all nodes in the variable's hierarchy.
-        col_widths = {}
-        row_data = []
         for node in nodes:
-            # Format the node.
-            col_data = ctx.new_variable(node, None, ...)._format_node(self, True)
-            row_data.append(self.update_widths(col_widths, col_data))
-
-            # Gather formatting data for each child node in the hierarchy.
-            for child in iter_nodes(node):
-                col_data = ctx.new_variable(child, None, ...)._format_node(self, False)
-                row_data.append(self.update_widths(col_widths, col_data))
+            format_node(node, self.var, True)
 
         # Sort the row data by path instead of by node ordering the hierarchy.
         if self.path_sort:
@@ -439,15 +493,15 @@ class StructureFormatter:
         else:
             raise TypeError(f'Unknown node type {ntype!r}.')
 
+        is_indirect = self._node.protocol is not None
         region = self._node.region
-        start = region.offset.absolute
-        end = region.offset.absolute + region.size - 1
 
         return {
             'type': type_,
-            'path': formatter.qualname(self._node, is_root),
-            'size': formatter.size(region.size),
-            'offset': formatter.offset_range(start, end),
+            'path': formatter.qualname(self._node, is_root) +
+                    (' [indirect view]' if is_indirect else ''),
+            'size': formatter.size(region),
+            'offset': formatter.offset_range(region),
             'data': {
                 'oid': region.oid,
                 'ordinal': region.ordinal,
@@ -464,16 +518,14 @@ class StructureFormatter:
 class ArrayFormatter:
     def _format_node(self, formatter, is_root=True):
         region = self._node.region
-        start = region.offset.absolute
-        end = region.offset.absolute + region.size - 1
         qualname = formatter.qualname(self._node, is_root)
         subscripts = ''.join(f'[:{f}]' for f in self._node.indexer.fields)
 
         return {
             'type': 'Array',
             'path': f'{qualname}{subscripts}',
-            'size': formatter.size(region.size),
-            'offset': formatter.offset_range(start, end),
+            'size': formatter.size(region),
+            'offset': formatter.offset_range(region),
             'data': {
                 'oid': region.oid,
                 'ordinal': region.ordinal,
@@ -497,8 +549,8 @@ class RegisterFormatter:
             'type': 'Register',
             'access': self._node.config.access.name,
             'path': formatter.qualname(self._node, is_root),
-            'size': formatter.size(region.size),
-            'offset': formatter.offset(region.offset.absolute),
+            'size': formatter.size(region),
+            'offset': formatter.offset(region),
             'value_hex': formatter.value_hex(value, region),
             'data': {
                 'oid': region.oid,
